+++ test-module/src/test_basics.rs
@@ -74,3 +74,10 @@ fn u64_overflow() -> Result<u64> {
 fn ignore_args(_: &Env, _: u8, _: u16) -> Result<()> {
     Ok(())
 }
+
+#[defun(mod_in_name = false)]
+fn copy_string_contents(v: Value, size: usize) -> Result<String> {
+    let mut buffer = vec![0u8; size];
+    let s = v.copy_string_contents(&mut buffer)?;
+    Ok(String::from_utf8_lossy(s).to_string())
+}